% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-server-module.R
\name{addServerModule}
\alias{addServerModule}
\title{Add a server-side only module}
\usage{
addServerModule(name)
}
\arguments{
\item{name}{Name of the server side module}
}
\value{
An invisible `TRUE` if successful
}
\description{
Add a server-side only module
}
