% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tab.R
\name{addTab}
\alias{addTab}
\title{Create a new tab}
\usage{
addTab(name, parent = "ui", tabs = "ui/body/tabs", active = TRUE,
  open = TRUE)
}
\arguments{
\item{name}{A character string. What to call the tab? Cannot start with a number.}

\item{parent}{A character string. The name of the tab the new parent should go under. Defaults to 'ui'
which indicates the tab should be accessible on the top level of the main menu}

\item{tabs}{A file path. The location of the top of the tabs in an app's ui directory}

\item{active}{A Boolean. Should the .yaml be initialized to immediately include the tab in the app?}

\item{open}{A Boolean. Should RStudio attempt to open the tab you just created?}
}
\description{
Create a new tab
}
