% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-file-naming-conventions.R
\name{checkFileNamingConventions}
\alias{checkFileNamingConventions}
\title{Check the naming conventions of the app files}
\usage{
checkFileNamingConventions(criteria = "[^a-z_\\\\.\\\\/]")
}
\arguments{
\item{criteria}{A regex expression to definite the allowable characters}
}
\description{
Check the naming conventions of the app files
}
