% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-tab.R
\name{deleteTab}
\alias{deleteTab}
\title{Remove an existing tab}
\usage{
deleteTab(name, cascade = FALSE)
}
\arguments{
\item{name}{A character string. What to call the tab? Cannot start with a number.}

\item{cascade}{A boolean, should it delete all tabs below it?}
}
\description{
Remove an existing tab
}
