% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-yaml.R
\name{evaluateYAML}
\alias{evaluateYAML}
\title{Evaluate a YAML list in place}
\usage{
evaluateYAML(yaml_list)
}
\arguments{
\item{yaml_list}{A possibly nested list of objects read from a YAML file}
}
\value{
A list
}
\description{
Evaluate a YAML list in place
}
