% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-tab-name.R
\name{extractTabName}
\alias{extractTabName}
\title{Extract the tab name from the file name}
\usage{
extractTabName(path)
}
\arguments{
\item{path}{An .R or .yaml file path}
}
\value{
A character string
}
\description{
Extract the tab name from the file name
}
\examples{
extractTabName("a.dot.delimited.name.tabname.R")
}
