% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-in-tree-structure.R
\name{filesInTreeStructure}
\alias{filesInTreeStructure}
\title{Determine a directory file structure}
\usage{
filesInTreeStructure(file, pattern = NULL)
}
\arguments{
\item{file}{The root directory or file}

\item{pattern}{An optional pattern to filter out directories containing files}
}
\value{
A list
}
\description{
Determine a directory file structure
}
