% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-a-tab.R
\name{findATab}
\alias{findATab}
\title{Find an existing tab}
\usage{
findATab(name, tabs = "ui/body/tabs")
}
\arguments{
\item{name}{The name of the tab to find the location for}

\item{tabs}{Where to start looking}
}
\value{
A filepath of the location of that tab
}
\description{
Find an existing tab
}
