% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-child-tabs.R
\name{getChildTabs}
\alias{getChildTabs}
\title{Get the names of the tabs that are nested under a given tab}
\usage{
getChildTabs(name)
}
\arguments{
\item{name}{A character string, the name of the tab}
}
\value{
A character vector
}
\description{
Get the names of the tabs that are nested under a given tab
}
