% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-existing-tab-names.R
\name{getExistingTabs}
\alias{getExistingTabs}
\title{Find the existing tab names}
\usage{
getExistingTabs(path = "ui/body/tabs")
}
\arguments{
\item{path}{Where to start the search}
}
\value{
A character vector
}
\description{
Find the existing tab names
}
\examples{
demo_app_ui <- system.file("extdata", "demo-app", "ui", package = "shinytabconstructor", mustWork = TRUE)
getExistingTabs(demo_app_ui)
}
