% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-locs.R
\name{getServerLocation}
\alias{getServerLocation}
\title{Find the physical location of the server file within project}
\usage{
getServerLocation()
}
\value{
A character string: The fully qualified file path to the server folder
}
\description{
Find the physical location of the server file within project
}
