% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-locs.R
\name{getTabLocation}
\alias{getTabLocation}
\title{Find the physical location of the tabs within the project}
\usage{
getTabLocation()
}
\value{
A character string: The fully qualified file path to the tabs folder
}
\description{
Find the physical location of the tabs within the project
}
