% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-menu.R
\name{makeMenu}
\alias{makeMenu}
\title{Make a menu from a configured hierarchical list}
\usage{
makeMenu(.list)
}
\arguments{
\item{.list}{A (possibly) nested list of the .yaml files defining a menu structure}
}
\value{
A list of html menu objects
}
\description{
Make a menu from a configured hierarchical list
}
