% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu-item-list.R
\name{menuItemList}
\alias{menuItemList}
\title{Create a menu item}
\usage{
menuItemList(text, .list, icon = NULL, badgeLabel = NULL,
  badgeColor = "green", tabName = NULL, href = NULL, newtab = TRUE,
  selected = NULL, expandedName = as.character(gsub("[[:space:]]", "",
  text)), startExpanded = FALSE)
}
\arguments{
\item{text}{Text to show for the menu item.}

\item{.list}{An optional list of \code{\link[shinydashboard]{menuSubItem}}s}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}. If
\code{NULL}, don't display an icon.}

\item{badgeLabel}{A label for an optional badge. Usually a number or a short
word like "new".}

\item{badgeColor}{A color for the badge. Valid colors are listed in
\link[shinydashboard]{validColors}.}

\item{tabName}{The name of a tab that this menu item will activate. Not
compatible with \code{href}.}

\item{href}{An link address. Not compatible with \code{tabName}.}

\item{newtab}{If \code{href} is supplied, should the link open in a new
browser tab?}

\item{selected}{If \code{TRUE}, this \code{menuItem} or \code{menuSubItem}
will start selected. If no item have \code{selected=TRUE}, then the first
\code{menuItem} will start selected.}

\item{expandedName}{A unique name given to each \code{menuItem} that serves
to indicate which one (if any) is currently expanded. (This is only applicable
to \code{menuItem}s that have children and it is mostly only useful for
bookmarking state.)}

\item{startExpanded}{Should this \code{menuItem} be expanded on app startup?
(This is only applicable to \code{menuItem}s that have children, and only
one of these can be expanded at any given time).}

\item{id}{For \code{sidebarMenu}, if \code{id} is present, this id will be
used for a Shiny input value, and it will report which tab is selected. For
example, if \code{id="tabs"}, then \code{input$tabs} will be the
\code{tabName} of the currently-selected tab. If you want to be able to
bookmark and restore the selected tab, an \code{id} is required.}
}
\description{
Create a menu item
}
\details{
Exactly the same as \code{\link[shinydashboard]{menuItem}} except that
    the \code{\link[shinydashboard]{menuSubItem}}s can be passed in a list.
    All creational credit to the maintainers of that package
}
