% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-item-list.R
\name{tabItemList}
\alias{tabItemList}
\title{Form a tab structure from one or more tab items}
\usage{
tabItemList(items)
}
\arguments{
\item{items}{One or more \code{\link[shinydashboard]{tabItem}} in a list}
}
\value{
A html div
}
\description{
Form a tab structure from one or more tab items
}
