% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tab.R
\name{addTab}
\alias{addTab}
\title{Create a new tab}
\usage{
addTab(name, parent = "ui", tabs = "ui/body/tabs", active = TRUE)
}
\arguments{
\item{name}{What to call the tab? Cannot start with a number.}

\item{parent}{The name of the tab the new parent should go under. Defaults to 'ui'
which indicates the tab should be accessible on the top level of the main menu}
}
\value{
\code{TRUE}
}
\description{
Create a new tab
}
