% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-duplicate-tab-names.R
\name{checkDuplicateTabNames}
\alias{checkDuplicateTabNames}
\title{Check the files against naming conventions}
\usage{
checkDuplicateTabNames(path = "ui/body/tabs/")
}
\arguments{
\item{path}{Root path}
}
\description{
Check the files against naming conventions
}
