% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-naming-conventions.R
\name{checkNamingConventions}
\alias{checkNamingConventions}
\title{Check the naming conventions of the app files}
\usage{
checkNamingConventions(files, criteria = "[^a-z0-9_\\\\.\\\\/]")
}
\arguments{
\item{files}{A character vector of file names}

\item{criteria}{A regex expression to definite the allowable characters}
}
\value{
\code{TRUE}
}
\description{
Check the naming conventions of the app files
}
\examples{
checkNamingConventions(c("01-desc.R", "a.good.name.R", "another_acceptable.name.R"))
}
