% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-tab.R
\name{findTab}
\alias{findTab}
\title{Find an existing tab}
\usage{
findTab(name)
}
\arguments{
\item{name}{The name of the tab to find the location for}
}
\value{
A filepath of the location of that tab
}
\description{
Find an existing tab
}
